Fix tests (Move `OperationSpecification` to `.operation`)
