+++ neoclient/middleware.py
@@ -2,10 +2,11 @@ from dataclasses import dataclass, field
 from typing import Optional, Protocol, Sequence
 
 import mediate
-from httpx import Request, Response
+from httpx import Response
 
 from .enums import HeaderName
 from .errors import ExpectedHeaderError, ExpectedStatusCodeError
+from .models import Request
 
 __all__: Sequence[str] = (
     "Middleware",
